+++ planetclicker/ui/panel.py
@@ -1,14 +1,15 @@
-from src.ui.uielement import UIElement
+from planetclicker.ui.uielement import UIElement
+
 
 #   panel element, just a base ui element that has children, updates those children as well
 class Panel(UIElement):
     def __init__(self, pos, width_height, color=None, label=""):
         super().__init__(pos, width_height, color, label)
         self.children = []
-        
+
     def add_child(self, to_add):
         self.children.append(to_add)
-        
+
     #   update children input
     def handle_input(self, events, pressed_keys):
         for child in self.children:
@@ -18,7 +19,7 @@ class Panel(UIElement):
     def update(self):
         for child in self.children:
             child.update()
-            
+
     #   draw children
     def draw(self, surface):
         super().draw(surface)
